commit de blocage a restaurer
