+++ gestionstocks/urls.py
@@ -4,5 +4,6 @@ from django.urls import include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('base.urls'))
+    path('', include('base.urls')),
+    path('matiere/', include('matiere.urls')),
 ]
