+++ templates/navbar.html
@@ -3,6 +3,7 @@
 >
   <a class="font-bold" href="/">GSA</a>
   <span class="flex justify-between gap-4">
+    <a class="font-light" href="{% url 'matiere' %}">Matiere Premières</a>
     <a class="font-light" href="/produit">Produits</a>
     <a class="font-light" href="/appro">Approvisionnement</a>
   </span>
